fix(editor): Move projects and rbac files
