+++ packages/editor-ui/src/features/projects/projects.store.ts
@@ -11,8 +11,8 @@ import type {
 	ProjectsCount,
 } from '@/features/projects/projects.types';
 import { useSettingsStore } from '@/stores/settings.store';
-import { hasPermission } from '@/rbac/permissions';
-import { ProjectTypes } from './projects.utils';
+import { hasPermission } from '@/utils/rbac/permissions';
+import { ProjectTypes } from '@/features/projects/projects.utils';
 import type { IWorkflowDb } from '@/Interface';
 
 export const useProjectsStore = defineStore('projects', () => {
