+++ packages/editor-ui/src/utils/rbac/checks/hasScope.ts
