+++ packages/editor-ui/src/utils/rbac/checks/isAuthenticated.ts
