+++ packages/editor-ui/src/utils/rbac/checks/isDefaultUser.ts
