+++ packages/editor-ui/src/utils/rbac/checks/isGuest.ts
