+++ packages/editor-ui/src/utils/rbac/checks/isValid.test.ts
@@ -1,4 +1,4 @@
-import { isValid } from '@/rbac/checks/isValid';
+import { isValid } from '@/utils/rbac/checks/isValid';
 
 describe('Checks', () => {
 	describe('isValid()', () => {
