+++ packages/editor-ui/src/utils/rbac/middleware/defaultUser.ts
@@ -1,7 +1,7 @@
 import type { RouterMiddleware } from '@/types/router';
 import { VIEWS } from '@/constants';
 import type { DefaultUserMiddlewareOptions } from '@/types/rbac';
-import { isDefaultUser } from '@/rbac/checks';
+import { isDefaultUser } from '@/utils/rbac/checks';
 
 export const defaultUserMiddleware: RouterMiddleware<DefaultUserMiddlewareOptions> = async (
 	_to,
